Fix unity build failure
