+++ Src/Demos/02_Triangle/TriangleDemo.cpp
@@ -13,23 +13,6 @@
 #include "Demos/DemoRenderUtils.h"
 #include "Render/Shader/ShaderParameterWriter.h"
 
-#if defined(GLAB_BACKEND_VULKAN) || defined(GLAB_BACKEND_METAL)
-namespace
-{
-struct ColoredVertex
-{
-    Math::Vec4 m_Position;
-    Math::Vec4 m_Color;
-};
-
-ShaderCompileRequest BuildShaderCompileRequest()
-{
-    return DemoRenderUtils::BuildGraphicsShaderCompileRequest(FileSystem::GetRootPath() / "Project" / "Shaders" /
-                                                              "DemoColored.slang");
-}
-} // namespace
-#endif
-
 TriangleDemo::TriangleDemo(uint32_t width, uint32_t height) : m_ViewportWidth(width), m_ViewportHeight(height) {}
 
 void TriangleDemo::OnAttach()
@@ -121,6 +104,7 @@ void TriangleDemo::CreateTriangleResources()
     Application& app = Application::Get();
     Device& device = app.GetDevice();
 
+    using DemoRenderUtils::ColoredVertex;
     static const std::array<ColoredVertex, 3> kVertices = {{
         {{0.0f, -0.65f, 0.0f, 1.0f}, {1.0f, 0.20f, 0.20f, 1.0f}},
         {{0.65f, 0.55f, 0.0f, 1.0f}, {0.20f, 0.90f, 0.35f, 1.0f}},
@@ -146,8 +130,10 @@ void TriangleDemo::CreateTriangleResources()
                                             m_IndexUploadBuffer);
     m_GeometryUploadPending = true;
 
-    m_ShaderProgram = device.CreateShaderProgram(
-        DemoRenderUtils::CompileShaderProgramDesc(BuildShaderCompileRequest(), "TriangleDemo"));
+    m_ShaderProgram = device.CreateShaderProgram(DemoRenderUtils::CompileShaderProgramDesc(
+        DemoRenderUtils::BuildGraphicsShaderCompileRequest(FileSystem::GetRootPath() / "Project" / "Shaders" /
+                                                           "DemoColored.slang"),
+        "TriangleDemo"));
     m_PipelineLayout = device.CreatePipelineLayout(m_ShaderProgram->DerivePipelineLayoutDesc());
     const PipelineLayoutDesc& pipelineLayoutDesc = m_PipelineLayout->GetDesc();
     m_FrameSetIndex = DemoRenderUtils::FindRequiredSetIndex(pipelineLayoutDesc, "gFrame", "TriangleDemo");
