+++ Src/Demos/06_TexturedRotatingCube/TexturedRotatingCubeDemo.cpp
@@ -13,32 +13,6 @@
 #include "Demos/DemoRenderUtils.h"
 #include "Render/Shader/ShaderParameterWriter.h"
 
-#if defined(GLAB_BACKEND_VULKAN) || defined(GLAB_BACKEND_METAL)
-namespace
-{
-struct TexturedVertex
-{
-    Math::Vec4 m_Position;
-    Math::Vec2 m_UV;
-};
-
-ShaderCompileRequest BuildShaderCompileRequest()
-{
-    return DemoRenderUtils::BuildGraphicsShaderCompileRequest(FileSystem::GetRootPath() / "Project" / "Shaders" /
-                                                              "DemoTextured.slang");
-}
-
-Math::Mat4 BuildViewProjection(uint32_t width, uint32_t height)
-{
-    const float aspect = height == 0 ? 1.0f : static_cast<float>(width) / static_cast<float>(height);
-    const Math::Mat4 projection = Math::PerspectiveRH_ZO(Math::Radians(55.0f), aspect, 0.1f, 100.0f);
-    const Math::Mat4 view =
-        Math::LookAt(Math::Vec3(0.0f, 0.0f, 3.1f), Math::Vec3(0.0f, 0.0f, 0.0f), Math::Vec3(0.0f, 1.0f, 0.0f));
-    return projection * view;
-}
-} // namespace
-#endif
-
 TexturedRotatingCubeDemo::TexturedRotatingCubeDemo(uint32_t width, uint32_t height)
     : m_ViewportWidth(width), m_ViewportHeight(height)
 {
@@ -150,8 +124,9 @@ void TexturedRotatingCubeDemo::OnResize(uint32_t width, uint32_t height)
     if (m_FrameSet != nullptr)
     {
         ShaderParameterWriter parameterWriter(m_ShaderProgram->GetReflection());
-        parameterWriter.SetMatrix4x4(
-            *m_FrameSet, "gFrame.viewProj", BuildViewProjection(m_ViewportWidth, m_ViewportHeight));
+        parameterWriter.SetMatrix4x4(*m_FrameSet,
+                                     "gFrame.viewProj",
+                                     DemoRenderUtils::BuildOrbitViewProjection(m_ViewportWidth, m_ViewportHeight));
     }
 
     if (m_DepthTexture != nullptr)
@@ -165,6 +140,7 @@ void TexturedRotatingCubeDemo::CreateCubeResources()
     Application& app = Application::Get();
     Device& device = app.GetDevice();
 
+    using DemoRenderUtils::TexturedVertex;
     static const std::array<TexturedVertex, 24> kVertices = {{
         {{-0.6f, -0.6f, 0.6f, 1.0f}, {0.0f, 1.0f}},  {{0.6f, -0.6f, 0.6f, 1.0f}, {1.0f, 1.0f}},
         {{0.6f, 0.6f, 0.6f, 1.0f}, {1.0f, 0.0f}},    {{-0.6f, 0.6f, 0.6f, 1.0f}, {0.0f, 0.0f}},
@@ -249,8 +225,10 @@ void TexturedRotatingCubeDemo::CreateCubeResources()
                        static_cast<uint64_t>(loadedImage.m_Pixels.size()));
     m_TextureUploadPending = true;
 
-    m_ShaderProgram = device.CreateShaderProgram(
-        DemoRenderUtils::CompileShaderProgramDesc(BuildShaderCompileRequest(), "TexturedRotatingCube"));
+    m_ShaderProgram = device.CreateShaderProgram(DemoRenderUtils::CompileShaderProgramDesc(
+        DemoRenderUtils::BuildGraphicsShaderCompileRequest(FileSystem::GetRootPath() / "Project" / "Shaders" /
+                                                           "DemoTextured.slang"),
+        "TexturedRotatingCube"));
     m_PipelineLayout = device.CreatePipelineLayout(m_ShaderProgram->DerivePipelineLayoutDesc());
     const PipelineLayoutDesc& pipelineLayoutDesc = m_PipelineLayout->GetDesc();
     m_FrameSetIndex = DemoRenderUtils::FindRequiredSetIndex(pipelineLayoutDesc, "gFrame", "TexturedRotatingCube");
@@ -379,7 +357,7 @@ void TexturedRotatingCubeDemo::UpdateShaderParameters()
 
     ShaderParameterWriter parameterWriter(m_ShaderProgram->GetReflection());
     parameterWriter.SetMatrix4x4(
-        *m_FrameSet, "gFrame.viewProj", BuildViewProjection(m_ViewportWidth, m_ViewportHeight));
+        *m_FrameSet, "gFrame.viewProj", DemoRenderUtils::BuildOrbitViewProjection(m_ViewportWidth, m_ViewportHeight));
     parameterWriter.SetFloat4(*m_FrameSet, "gFrame.tint", Math::Vec4(1.0f, 1.0f, 1.0f, 1.0f));
     parameterWriter.SetFloat(*m_FrameSet, "gFrame.time", m_RotationSeconds);
     parameterWriter.SetFloat4(*m_MaterialSet, "gMaterial.baseColor", Math::Vec4(1.0f, 1.0f, 1.0f, 1.0f));
